moved docs build to prestart
