+++ package.json
@@ -6,6 +6,7 @@
     "netsblox": "./bin/netsblox"
   },
   "scripts": {
+    "prestart": "node bin/prestart.js",
     "start": "node bin/netsblox-start.js",
     "dev": "npx nodemon --ignore src/browser/dist -w src -w test --exec 'npm run postinstall && npm start'",
     "testmon": "npx nodemon -w src -w test --exec 'npm run test'",
